feature: API호출, 데이터 화면에 출력 (#116)
